move filtering rules to core for use in formplayer
